remove modules.py and make examples independent
